# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance tracking: transactions, categories, dashboard and cash flow reports.",
        "title": "Financial Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Dashboard summary",
                "parameters": [
                    {
                        "type": "string",
                        "default": "month",
                        "description": "today, week, month, year or custom",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Required when range is custom",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Required when range is custom",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.DashboardSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/cash-flow": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Cash flow report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "INCOME, EXPENSE or TRANSFER",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.CashFlowResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/cash-flow/excel": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Cash flow report as Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "INCOME, EXPENSE or TRANSFER",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/cash-flow/pdf": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Cash flow report as PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "INCOME, EXPENSE or TRANSFER",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/daily-cash-flow": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "The same report the Telegram bot sends, as JSON. Sections with nothing to say are\nomitted, and a wallet or bank only reports a variance once a balance has been\nrecorded for it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Daily cash flow report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report date, YYYY-MM-DD. Defaults to today.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.DailyCashFlowReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/expense-by-category": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Expense by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.ExpenseByCategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Monthly report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "INCOME, EXPENSE or TRANSFER",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.MonthlyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Period summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date, YYYY-MM-DD",
                        "name": "date_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date, YYYY-MM-DD",
                        "name": "date_to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "INCOME, EXPENSE or TRANSFER",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.SummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.CashFlowResponse": {
            "type": "object",
            "properties": {
                "net_cash_flow": {
                    "type": "string",
                    "example": "6500000.00"
                },
                "period": {
                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.Period"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.CashFlowRow"
                    }
                },
                "total_expense": {
                    "type": "string",
                    "example": "8500000.00"
                },
                "total_income": {
                    "type": "string",
                    "example": "15000000.00"
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.DashboardSummaryResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string",
                    "example": "6500000.00"
                },
                "cash_flow_trend": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.CashFlowPoint"
                    }
                },
                "expense_by_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.ExpenseByCategoryRow"
                    }
                },
                "income_vs_expense": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.IncomeVsExpensePoint"
                    }
                },
                "period": {
                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.Period"
                },
                "recent_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.TransactionResponse"
                    }
                },
                "total_expense": {
                    "type": "string",
                    "example": "8500000.00"
                },
                "total_income": {
                    "type": "string",
                    "example": "15000000.00"
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 125
                }
            }
        },
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.ExpenseByCategoryResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.Period"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.ExpenseByCategoryRow"
                    }
                },
                "total": {
                    "type": "string",
                    "example": "8500000.00"
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.IncomeVsExpensePoint": {
            "type": "object",
            "properties": {
                "expense": {
                    "type": "string",
                    "example": "8500000.00"
                },
                "income": {
                    "type": "string",
                    "example": "15000000.00"
                },
                "label": {
                    "type": "string",
                    "example": "2026-09"
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.MonthlyResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.Period"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.MonthlyRow"
                    }
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.SummaryResponse": {
            "type": "object",
            "properties": {
                "net_cash_flow": {
                    "type": "string",
                    "example": "6500000.00"
                },
                "period": {
                    "$ref": "#/definitions/github_com_rezanii_tracking-cashflow_apps_backend_internal_dto.Period"
                },
                "total_expense": {
                    "type": "string",
                    "example": "8500000.00"
                },
                "total_income": {
                    "type": "string",
                    "example": "15000000.00"
                }
            }
        },
...
        "github_com_rezanii_tracking-cashflow_apps_backend_internal_utils.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Send the access token as: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}